chore: 불필요한 파일 제거 및 Test용 코드 위치 변경
