--- src/components/domain/post/ViewPost.css
@@ -1,54 +1,0 @@
-.PostCard-Container {
-  display: flex;
-  justify-content: flex-start;
-  align-items: center;
-
-  background-color: beige;
-  width: 275px;
-  height: 260px;
-
-  border: 1px solid beige;
-  border-radius: 16px;
-
-  .PostCard-Form {
-    padding: 30px 24px 20px;
-  }
-
-  .PostCard-Content {
-    display: flex;
-    flex-direction: column;
-    margin-bottom: 43px;
-
-    gap: 12px;
-  }
-
-  .PostCard-Content .Name {
-    color: #181818;
-    font-size: 24px;
-    font-weight: 700;
-  }
-
-  .PostCard-Content .messageCount {
-    color: #3a3a3a;
-    font-size: 16px;
-    font-weight: 400;
-  }
-
-  .Reactions-Form {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-
-    border-top: 1px solid grey;
-    padding-top: 17px;
-
-    gap: 10px;
-  }
-
-  .Reactions-Form .Reaction-Content {
-    width: fit-content;
-    background-color: grey;
-    border-radius: 32px;
-    padding: 8px 12px;
-  }
-}
